refactor: move code into lib/services folder and write spec
