+++ .gitignore
@@ -45,3 +45,5 @@ public/uploads/
 /yarn-error.log
 yarn-debug.log*
 .yarn-integrity
+
+.DS_store
